add & move: folder components, app to components
