+++ src/components/App/App.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import logo from './logo.svg';
+import logo from '../../logo.svg';
+
 import './App.css';
 
-function App() {
+const App: React.FC = () => {
   return (
     <div className="App">
       <header className="App-header">
@@ -23,4 +23,4 @@ function App() {
   );
 }
 
-export default App;
+export { App };
